Cube faces can now be rotated
This rotation method is primarily for testing if GetPiecesWithinArea is working properly for pieces that came from other faces. I plan to replace the algorythm for rotating to something faster that doesn't rely on parenting back and forth. For now, this will remain mostly untouched until I get the VR interaction in.